# Conflicts:
#	app/Http/Controllers/CartController.php
#	app/Http/routes.php

+++ app/Http/Controllers/Admin/AdminController.php
@@ -10,12 +10,14 @@ use App\Http\Controllers\Controller;
 
 /**
  * Class AdminController
+ * @author liuzhiqi
  * @package App\Http\Controller\Admin
  */
 class AdminController extends Controller
 {
     /**
      * Display a listing of the resource.
+     * @author liuzhiqi
      *
      * @return \Illuminate\Http\Response
      */
@@ -27,16 +29,19 @@ class AdminController extends Controller
 
     /**
      * Show the form for creating a new resource.
+     * @author liuzhiqi
      *
      * @return \Illuminate\Http\Response
      */
     public function create()
     {
         //
+        return view('admin.admins.create');
     }
 
     /**
      * Store a newly created resource in storage.
+     * @author liuzhiqi
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
@@ -49,6 +54,7 @@ class AdminController extends Controller
     /**
      * Display the specified resource.
      * 管理员信息展示页面
+     * @author liuzhiqi
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
@@ -68,6 +74,7 @@ class AdminController extends Controller
 
     /**
      * Show the form for editing the specified resource.
+     * @author liuzhiqi
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
@@ -87,6 +94,7 @@ class AdminController extends Controller
     /**
      * Update the specified resource in storage.
      * 修改管理员信息
+     * @author liuzhiqi
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
@@ -103,15 +111,19 @@ class AdminController extends Controller
     /**
      * Remove the specified resource from storage.
      * 删除管理员
+     * @author liuzhiqi
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        if (AdminUser::destroy($id)) {
-            return redirect('admin/admins')->with(['msg' => '删除成功！！！']);
+        if( \Auth::guard('admin')->user()->type == 0) {
+            if (AdminUser::destroy($id)) {
+                return redirect('admin/admins')->with(['msg' => '删除成功！！！']);
+            }
         }
+        return back()->with(['msg' => '权限不足']);
     }
 
 }
